requirements.txt
Django==5.1.1
python-decouple==3.8
django-widget-tweaks==1.5.0
django-bootstrap-form==3.4
xlrd==2.0.1
xlwt==1.3.0
xhtml2pdf==0.2.16
django-extensions==3.2.3
django-ninja==1.3.0
